Moving namespace query to types so it is accessible for tests
