+++ apps/api-catalogue/components/ServiceCard/ServiceCard.tsx
@@ -8,7 +8,7 @@ import { useIsomorphicLayoutEffect, useWindowSize } from 'react-use'
 import { theme } from '@island.is/island-ui/theme'
 import { capitalize } from '../../utils'
 
-import { GetNamespaceQuery } from '../../graphql/schema'
+import { GetNamespaceQuery } from '../../types'
 import { useNamespace } from '../../hooks'
 
 export interface ServiceCardProps {
