+++ apps/api-catalogue/components/ServiceFilter/ServiceFilter.tsx
@@ -11,7 +11,7 @@ import {
   TypeCategory,
 } from '@island.is/api-catalogue/consts'
 
-import { GetNamespaceQuery } from '../../graphql/schema'
+import { GetNamespaceQuery } from '../../types'
 import { useNamespace } from '../../hooks'
 
 type IconVariantTypes = 'default' | 'sidebar'
