+++ apps/api-catalogue/screens/DesignGuide/DesignGuide.tsx
@@ -11,9 +11,8 @@ import {
 
 import { useNamespace } from '../../hooks'
 import { QueryGetNamespaceArgs } from '@island.is/api/schema'
-import { GetNamespaceQuery } from '../../graphql/schema'
 import { GET_NAMESPACE_QUERY } from '../Queries'
-import { Screen } from '../../types'
+import { Screen, GetNamespaceQuery } from '../../types'
 import initApollo from '../../graphql/client'
 
 import cn from 'classnames'
