+++ apps/api-catalogue/screens/Home/Home.tsx
@@ -10,9 +10,8 @@ import cn from 'classnames'
 
 import { useNamespace } from '../../hooks'
 import { QueryGetNamespaceArgs } from '@island.is/api/schema'
-import { GetNamespaceQuery } from '../../graphql/schema'
 import { GET_NAMESPACE_QUERY } from '../Queries'
-import { Screen } from '../../types'
+import { Screen, GetNamespaceQuery } from '../../types'
 import initApollo from '../../graphql/client'
 
 interface HomeProps {
