+++ apps/api-catalogue/types.d.ts
@@ -1,5 +1,9 @@
 import { NextComponentType } from 'next'
 import { NextPageContext } from 'next/dist/next-server/lib/utils'
+import { 
+  Maybe,
+  Namespace
+ } from '@island.is/api/schema'
 
 export type GetInitialPropsContext<Context> = Context & {
   locale: string
@@ -10,3 +14,8 @@ export type Screen<Props = {}> = NextComponentType<
   Props,
   Props
 >
+
+/* GraphQL Types */
+export type GetNamespaceQuery = { __typename?: 'Query' } & {
+  getNamespace?: Maybe<{ __typename?: 'Namespace' } & Pick<Namespace, 'fields'>>
+}
